+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - **RSpec 3.5+ is requiered for RSpec features** ([@palkan][])
 
+- Add `with_logging` and `with_ar_logging` helpers to logging recipe. ([@palkan][])
+
 - Make `before_all` for Active Record `lock_thread` aware. ([@palkan][])
 
   `before_all` can went crazy if you open multiple connections within it